# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API ePortfolio",
      "description": "Daftar API ePortfolio",
      "version": "1.0"
    },
    "servers":[
        {
            "url": "http://127.0.0.1:8800/api",
            "description":"localhost"
        },
        {
            "url": "http://103.187.223.15:8800/api",
            "description":"dev"
        }
    ],
...
    "produces": ["application/json"],
    "paths": {
...
      "/activities":{
        "post":{
          "tags":[
            "Activity"
          ],
          "summary":"Create Activity",
          "description":"Create a new activity",
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "image":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "type":{
                      "type":"string",
                      "enum": ["academic", "non-academic"]
                    },
                    "startDate":{
                      "type":"string"
                    },
                    "endDate":{
                      "type":"string",
                      "nullable":true
                    },
                    "desc":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/activities/{id}":{
        "put":{
          "tags":[
            "Activity"
          ],
          "summary":"Update Activity",
          "description":"Update activity",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"activity id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "image":{
                      "type":"string"
                    },
                    "title":{
                      "type":"string"
                    },
                    "type":{
                      "type":"string",
                      "enum": ["academic", "non-academic"]
                    },
                    "startDate":{
                      "type":"string"
                    },
                    "endDate":{
                      "type":"string",
                      "nullable":true
                    },
                    "desc":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Activity not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "delete":{
          "tags":[
            "Activity"
          ],
          "summary":"Delete Activity",
          "description":"Delete activity",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"activity id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Activity not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "get":{
          "tags":[
            "Activity"
          ],
          "summary":"Get Activity",
          "description":"Get activity data",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"activity id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Activity not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/activities/all/{id}":{
        "get":{
          "tags":[
            "Activity"
          ],
          "summary":"Get All Activities From User",
          "description":"Get all activities data from user",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/activities/last/{id}":{
        "get":{
          "tags":[
            "Activity"
          ],
          "summary":"Get Last Activity From User",
          "description":"Get last activity data from user",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/activities/search/{keyword}":{
        "get":{
          "tags":[
            "Activity"
          ],
          "summary":"Get Activity by Keyword",
          "description":"Get activity data by keyword",
          "parameters":[
            {
              "in":"path",
              "name": "keyword",
              "description":"keyword"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
...
      "/album":{
        "post":{
          "tags":[
            "Album"
          ],
          "summary":"Upload image",
          "description":"Upload image to album",
          "consumes":"multipart/form-data",
          "requestBody":{
            "content":{
              "multipart/form-data":{
                "schema":{
                  "type": "object",
                  "properties":{
                    "fileAlbum":{
                      "type":"string",
                      "format":"binary"
                    },
                    "data":{
                      "type":"object",
                      "properties":{
                        "userId":{
                          "type":"string"
                        },
                        "filename":{
                          "type":"string"
                        },
                        "filesize":{
                          "type":"integer"
                        },
                        "type":{
                          "type":"string",
                          "enum": ["image", "video"]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/album/video":{
        "post":{
          "tags":[
            "Album"
          ],
          "summary":"Upload video",
          "description":"Upload video to album",
          "consumes":"multipart/form-data",
          "requestBody":{
            "content":{
              "multipart/form-data":{
                "schema":{
                  "type": "object",
                  "properties":{
                    "fileAlbum":{
                      "type":"string",
                      "format":"binary"
                    },
                    "data":{
                      "type":"object",
                      "properties":{
                        "userId":{
                          "type":"string"
                        },
                        "filename":{
                          "type":"string"
                        },
                        "filesize":{
                          "type":"integer"
                        },
                        "type":{
                          "type":"string",
                          "enum": ["image", "video"]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/album/{id}":{
        "delete":{
          "tags":[
            "Album"
          ],
          "summary":"Delete file",
          "description":"Delete file from album",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"file id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Project not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/album/all/{id}":{
        "get":{
          "tags":[
            "Album"
          ],
          "summary":"Get All Files From User Album",
          "description":"Get all files data from user album",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      }
    }
 }